Rename ConnectToInternetXxx to NetworkXxx
Ref: #1844
