+++ packages/ubuntu_desktop_installer/lib/installer.dart
@@ -393,8 +393,8 @@ class _UbuntuDesktopInstallerWizardState
           builder: KeyboardLayoutPage.create,
           userData: InstallationStep.keyboard.index,
         ),
-        Routes.connectToInternet: WizardRoute(
-          builder: ConnectToInternetPage.create,
+        Routes.network: WizardRoute(
+          builder: NetworkPage.create,
           userData: InstallationStep.network.index,
           onNext: (_) =>
               service.hasEnoughDiskSpace ? Routes.updatesOtherSoftware : null,
