+++ packages/ubuntu_desktop_installer/lib/pages/network/connect_model.dart
@@ -2,7 +2,7 @@ import 'package:safe_change_notifier/safe_change_notifier.dart';
 import 'package:ubuntu_logger/ubuntu_logger.dart';
 
 /// @internal
-final log = Logger('connect_to_internet');
+final log = Logger('network');
 
 enum ConnectMode {
   none,
