+++ packages/ubuntu_desktop_installer/lib/pages/network/ethernet_view.dart
