+++ packages/ubuntu_desktop_installer/lib/pages/network/hidden_wifi_view.dart
