+++ packages/ubuntu_desktop_installer/lib/pages/network/network_tile.dart
