+++ packages/ubuntu_desktop_installer/lib/pages/network/wifi_view.dart
