+++ packages/ubuntu_desktop_installer/test/network/device_model_test.mocks.dart
