+++ packages/ubuntu_desktop_installer/test/network/ethernet_view_test.mocks.dart
@@ -1,17 +1,15 @@
 // Mocks generated by Mockito 5.3.2 from annotations
-// in ubuntu_desktop_installer/test/connect_to_internet/hidden_wifi_view_test.dart.
+// in ubuntu_desktop_installer/test/network/ethernet_view_test.dart.
 // Do not manually edit this file.
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
-import 'dart:async' as _i6;
-import 'dart:ui' as _i7;
+import 'dart:async' as _i5;
+import 'dart:ui' as _i6;
 
 import 'package:mockito/mockito.dart' as _i1;
-import 'package:ubuntu_desktop_installer/pages/connect_to_internet/connect_model.dart'
-    as _i5;
-import 'package:ubuntu_desktop_installer/pages/connect_to_internet/hidden_wifi_model.dart'
+import 'package:ubuntu_desktop_installer/pages/network/connect_model.dart'
     as _i4;
-import 'package:ubuntu_desktop_installer/pages/connect_to_internet/wifi_model.dart'
+import 'package:ubuntu_desktop_installer/pages/network/ethernet_model.dart'
     as _i3;
 import 'package:ubuntu_desktop_installer/services.dart' as _i2;
 
@@ -37,8 +35,9 @@ class _FakeNetworkService_0 extends _i1.SmartFake
         );
 }
 
-class _FakeWifiDevice_1 extends _i1.SmartFake implements _i3.WifiDevice {
-  _FakeWifiDevice_1(
+class _FakeEthernetDevice_1 extends _i1.SmartFake
+    implements _i3.EthernetDevice {
+  _FakeEthernetDevice_1(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -58,11 +57,11 @@ class _FakeNetworkManagerDevice_2 extends _i1.SmartFake
         );
 }
 
-/// A class which mocks [HiddenWifiModel].
+/// A class which mocks [EthernetModel].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockHiddenWifiModel extends _i1.Mock implements _i4.HiddenWifiModel {
-  MockHiddenWifiModel() {
+class MockEthernetModel extends _i1.Mock implements _i3.EthernetModel {
+  MockEthernetModel() {
     _i1.throwOnMissingStub(this);
   }
 
@@ -77,13 +76,13 @@ class MockHiddenWifiModel extends _i1.Mock implements _i4.HiddenWifiModel {
         returnValue: false,
       ) as bool);
   @override
-  bool get hasActiveConnection => (super.noSuchMethod(
-        Invocation.getter(#hasActiveConnection),
+  bool get isConnecting => (super.noSuchMethod(
+        Invocation.getter(#isConnecting),
         returnValue: false,
       ) as bool);
   @override
-  bool get isConnecting => (super.noSuchMethod(
-        Invocation.getter(#isConnecting),
+  bool get hasActiveConnection => (super.noSuchMethod(
+        Invocation.getter(#hasActiveConnection),
         returnValue: false,
       ) as bool);
   @override
@@ -92,15 +91,10 @@ class MockHiddenWifiModel extends _i1.Mock implements _i4.HiddenWifiModel {
         returnValue: false,
       ) as bool);
   @override
-  _i5.ConnectMode get connectMode => (super.noSuchMethod(
+  _i4.ConnectMode get connectMode => (super.noSuchMethod(
         Invocation.getter(#connectMode),
-        returnValue: _i5.ConnectMode.none,
-      ) as _i5.ConnectMode);
-  @override
-  String get ssid => (super.noSuchMethod(
-        Invocation.getter(#ssid),
-        returnValue: '',
-      ) as String);
+        returnValue: _i4.ConnectMode.none,
+      ) as _i4.ConnectMode);
   @override
   _i2.NetworkService get service => (super.noSuchMethod(
         Invocation.getter(#service),
@@ -110,15 +104,15 @@ class MockHiddenWifiModel extends _i1.Mock implements _i4.HiddenWifiModel {
         ),
       ) as _i2.NetworkService);
   @override
-  List<_i3.WifiDevice> get devices => (super.noSuchMethod(
+  List<_i3.EthernetDevice> get devices => (super.noSuchMethod(
         Invocation.getter(#devices),
-        returnValue: <_i3.WifiDevice>[],
-      ) as List<_i3.WifiDevice>);
+        returnValue: <_i3.EthernetDevice>[],
+      ) as List<_i3.EthernetDevice>);
   @override
-  _i6.Stream<dynamic> get onAvailabilityChanged => (super.noSuchMethod(
+  _i5.Stream<dynamic> get onAvailabilityChanged => (super.noSuchMethod(
         Invocation.getter(#onAvailabilityChanged),
-        returnValue: _i6.Stream<dynamic>.empty(),
-      ) as _i6.Stream<dynamic>);
+        returnValue: _i5.Stream<dynamic>.empty(),
+      ) as _i5.Stream<dynamic>);
   @override
   bool get hasListeners => (super.noSuchMethod(
         Invocation.getter(#hasListeners),
@@ -130,30 +124,13 @@ class MockHiddenWifiModel extends _i1.Mock implements _i4.HiddenWifiModel {
         returnValue: false,
       ) as bool);
   @override
-  _i6.Future<dynamic> connect() => (super.noSuchMethod(
+  void onSelected() => super.noSuchMethod(
         Invocation.method(
-          #connect,
+          #onSelected,
           [],
-        ),
-        returnValue: _i6.Future<dynamic>.value(),
-      ) as _i6.Future<dynamic>);
-  @override
-  void setSsid(String? ssid) => super.noSuchMethod(
-        Invocation.method(
-          #setSsid,
-          [ssid],
         ),
         returnValueForMissingStub: null,
       );
-  @override
-  _i6.Future<void> enable() => (super.noSuchMethod(
-        Invocation.method(
-          #enable,
-          [],
-        ),
-        returnValue: _i6.Future<void>.value(),
-        returnValueForMissingStub: _i6.Future<void>.value(),
-      ) as _i6.Future<void>);
   @override
   void onDeselected() => super.noSuchMethod(
         Invocation.method(
@@ -163,13 +140,23 @@ class MockHiddenWifiModel extends _i1.Mock implements _i4.HiddenWifiModel {
         returnValueForMissingStub: null,
       );
   @override
-  void onSelected() => super.noSuchMethod(
+  _i5.Future<void> enable() => (super.noSuchMethod(
         Invocation.method(
-          #onSelected,
+          #enable,
           [],
         ),
-        returnValueForMissingStub: null,
-      );
+        returnValue: _i5.Future<void>.value(),
+        returnValueForMissingStub: _i5.Future<void>.value(),
+      ) as _i5.Future<void>);
+  @override
+  _i5.Future<void> connect() => (super.noSuchMethod(
+        Invocation.method(
+          #connect,
+          [],
+        ),
+        returnValue: _i5.Future<void>.value(),
+        returnValueForMissingStub: _i5.Future<void>.value(),
+      ) as _i5.Future<void>);
   @override
   List<_i2.NetworkManagerDevice> getDevices() => (super.noSuchMethod(
         Invocation.method(
@@ -179,38 +166,38 @@ class MockHiddenWifiModel extends _i1.Mock implements _i4.HiddenWifiModel {
         returnValue: <_i2.NetworkManagerDevice>[],
       ) as List<_i2.NetworkManagerDevice>);
   @override
-  _i3.WifiDevice createDevice(_i2.NetworkManagerDevice? device) =>
+  _i3.EthernetDevice createDevice(_i2.NetworkManagerDevice? device) =>
       (super.noSuchMethod(
         Invocation.method(
           #createDevice,
           [device],
         ),
-        returnValue: _FakeWifiDevice_1(
+        returnValue: _FakeEthernetDevice_1(
           this,
           Invocation.method(
             #createDevice,
             [device],
           ),
         ),
-      ) as _i3.WifiDevice);
+      ) as _i3.EthernetDevice);
   @override
-  _i6.Future<void> init() => (super.noSuchMethod(
+  _i5.Future<void> init() => (super.noSuchMethod(
         Invocation.method(
           #init,
           [],
         ),
-        returnValue: _i6.Future<void>.value(),
-        returnValueForMissingStub: _i6.Future<void>.value(),
-      ) as _i6.Future<void>);
+        returnValue: _i5.Future<void>.value(),
+        returnValueForMissingStub: _i5.Future<void>.value(),
+      ) as _i5.Future<void>);
   @override
-  _i6.Future<void> cleanup() => (super.noSuchMethod(
+  _i5.Future<void> cleanup() => (super.noSuchMethod(
         Invocation.method(
           #cleanup,
           [],
         ),
-        returnValue: _i6.Future<void>.value(),
-        returnValueForMissingStub: _i6.Future<void>.value(),
-      ) as _i6.Future<void>);
+        returnValue: _i5.Future<void>.value(),
+        returnValueForMissingStub: _i5.Future<void>.value(),
+      ) as _i5.Future<void>);
   @override
   void dispose() => super.noSuchMethod(
         Invocation.method(
@@ -228,7 +215,7 @@ class MockHiddenWifiModel extends _i1.Mock implements _i4.HiddenWifiModel {
         returnValueForMissingStub: null,
       );
   @override
-  bool isSelectedDevice(_i3.WifiDevice? device) => (super.noSuchMethod(
+  bool isSelectedDevice(_i3.EthernetDevice? device) => (super.noSuchMethod(
         Invocation.method(
           #isSelectedDevice,
           [device],
@@ -236,7 +223,7 @@ class MockHiddenWifiModel extends _i1.Mock implements _i4.HiddenWifiModel {
         returnValue: false,
       ) as bool);
   @override
-  void selectDevice(_i3.WifiDevice? device) => super.noSuchMethod(
+  void selectDevice(_i3.EthernetDevice? device) => super.noSuchMethod(
         Invocation.method(
           #selectDevice,
           [device],
@@ -252,7 +239,7 @@ class MockHiddenWifiModel extends _i1.Mock implements _i4.HiddenWifiModel {
         returnValueForMissingStub: null,
       );
   @override
-  void setProperties(_i6.Stream<List<String>>? properties) =>
+  void setProperties(_i5.Stream<List<String>>? properties) =>
       super.noSuchMethod(
         Invocation.method(
           #setProperties,
@@ -263,7 +250,7 @@ class MockHiddenWifiModel extends _i1.Mock implements _i4.HiddenWifiModel {
   @override
   void addPropertyListener(
     String? property,
-    _i7.VoidCallback? onChanged,
+    _i6.VoidCallback? onChanged,
   ) =>
       super.noSuchMethod(
         Invocation.method(
@@ -292,7 +279,7 @@ class MockHiddenWifiModel extends _i1.Mock implements _i4.HiddenWifiModel {
         returnValueForMissingStub: null,
       );
   @override
-  void addListener(_i7.VoidCallback? listener) => super.noSuchMethod(
+  void addListener(_i6.VoidCallback? listener) => super.noSuchMethod(
         Invocation.method(
           #addListener,
           [listener],
@@ -300,7 +287,7 @@ class MockHiddenWifiModel extends _i1.Mock implements _i4.HiddenWifiModel {
         returnValueForMissingStub: null,
       );
   @override
-  void removeListener(_i7.VoidCallback? listener) => super.noSuchMethod(
+  void removeListener(_i6.VoidCallback? listener) => super.noSuchMethod(
         Invocation.method(
           #removeListener,
           [listener],
@@ -309,34 +296,14 @@ class MockHiddenWifiModel extends _i1.Mock implements _i4.HiddenWifiModel {
       );
 }
 
-/// A class which mocks [WifiDevice].
+/// A class which mocks [EthernetDevice].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockWifiDevice extends _i1.Mock implements _i3.WifiDevice {
-  MockWifiDevice() {
+class MockEthernetDevice extends _i1.Mock implements _i3.EthernetDevice {
+  MockEthernetDevice() {
     _i1.throwOnMissingStub(this);
   }
 
-  @override
-  bool get isActive => (super.noSuchMethod(
-        Invocation.getter(#isActive),
-        returnValue: false,
-      ) as bool);
-  @override
-  List<_i3.AccessPoint> get accessPoints => (super.noSuchMethod(
-        Invocation.getter(#accessPoints),
-        returnValue: <_i3.AccessPoint>[],
-      ) as List<_i3.AccessPoint>);
-  @override
-  bool get scanning => (super.noSuchMethod(
-        Invocation.getter(#scanning),
-        returnValue: false,
-      ) as bool);
-  @override
-  int get lastScan => (super.noSuchMethod(
-        Invocation.getter(#lastScan),
-        returnValue: 0,
-      ) as int);
   @override
   _i2.NetworkManagerDevice get device => (super.noSuchMethod(
         Invocation.getter(#device),
@@ -355,6 +322,11 @@ class MockWifiDevice extends _i1.Mock implements _i3.WifiDevice {
         Invocation.getter(#state),
         returnValue: _i2.NetworkManagerDeviceState.unknown,
       ) as _i2.NetworkManagerDeviceState);
+  @override
+  bool get isActive => (super.noSuchMethod(
+        Invocation.getter(#isActive),
+        returnValue: false,
+      ) as bool);
   @override
   bool get isConnecting => (super.noSuchMethod(
         Invocation.getter(#isConnecting),
@@ -413,14 +385,6 @@ class MockWifiDevice extends _i1.Mock implements _i3.WifiDevice {
         returnValueForMissingStub: null,
       );
   @override
-  void dispose() => super.noSuchMethod(
-        Invocation.method(
-          #dispose,
-          [],
-        ),
-        returnValueForMissingStub: null,
-      );
-  @override
   void updateDevice(_i2.NetworkManagerDevice? device) => super.noSuchMethod(
         Invocation.method(
           #updateDevice,
@@ -429,57 +393,16 @@ class MockWifiDevice extends _i1.Mock implements _i3.WifiDevice {
         returnValueForMissingStub: null,
       );
   @override
-  bool isActiveAccessPoint(_i3.AccessPoint? accessPoint) => (super.noSuchMethod(
-        Invocation.method(
-          #isActiveAccessPoint,
-          [accessPoint],
-        ),
-        returnValue: false,
-      ) as bool);
-  @override
-  bool isSelectedAccessPoint(_i3.AccessPoint? accessPoint) =>
-      (super.noSuchMethod(
-        Invocation.method(
-          #isSelectedAccessPoint,
-          [accessPoint],
-        ),
-        returnValue: false,
-      ) as bool);
-  @override
-  void selectAccessPoint(_i3.AccessPoint? accessPoint) => super.noSuchMethod(
-        Invocation.method(
-          #selectAccessPoint,
-          [accessPoint],
-        ),
-        returnValueForMissingStub: null,
-      );
-  @override
-  _i6.Future<void> requestScan({String? ssid}) => (super.noSuchMethod(
-        Invocation.method(
-          #requestScan,
-          [],
-          {#ssid: ssid},
-        ),
-        returnValue: _i6.Future<void>.value(),
-        returnValueForMissingStub: _i6.Future<void>.value(),
-      ) as _i6.Future<void>);
-  @override
-  _i3.AccessPoint? findAccessPoint(String? ssid) =>
-      (super.noSuchMethod(Invocation.method(
-        #findAccessPoint,
-        [ssid],
-      )) as _i3.AccessPoint?);
-  @override
-  _i6.Future<void> disconnect() => (super.noSuchMethod(
+  _i5.Future<void> disconnect() => (super.noSuchMethod(
         Invocation.method(
           #disconnect,
           [],
         ),
-        returnValue: _i6.Future<void>.value(),
-        returnValueForMissingStub: _i6.Future<void>.value(),
-      ) as _i6.Future<void>);
+        returnValue: _i5.Future<void>.value(),
+        returnValueForMissingStub: _i5.Future<void>.value(),
+      ) as _i5.Future<void>);
   @override
-  void setProperties(_i6.Stream<List<String>>? properties) =>
+  void setProperties(_i5.Stream<List<String>>? properties) =>
       super.noSuchMethod(
         Invocation.method(
           #setProperties,
@@ -490,7 +413,7 @@ class MockWifiDevice extends _i1.Mock implements _i3.WifiDevice {
   @override
   void addPropertyListener(
     String? property,
-    _i7.VoidCallback? onChanged,
+    _i6.VoidCallback? onChanged,
   ) =>
       super.noSuchMethod(
         Invocation.method(
@@ -519,7 +442,15 @@ class MockWifiDevice extends _i1.Mock implements _i3.WifiDevice {
         returnValueForMissingStub: null,
       );
   @override
-  void addListener(_i7.VoidCallback? listener) => super.noSuchMethod(
+  void dispose() => super.noSuchMethod(
+        Invocation.method(
+          #dispose,
+          [],
+        ),
+        returnValueForMissingStub: null,
+      );
+  @override
+  void addListener(_i6.VoidCallback? listener) => super.noSuchMethod(
         Invocation.method(
           #addListener,
           [listener],
@@ -527,7 +458,7 @@ class MockWifiDevice extends _i1.Mock implements _i3.WifiDevice {
         returnValueForMissingStub: null,
       );
   @override
-  void removeListener(_i7.VoidCallback? listener) => super.noSuchMethod(
+  void removeListener(_i6.VoidCallback? listener) => super.noSuchMethod(
         Invocation.method(
           #removeListener,
           [listener],
