+++ packages/ubuntu_desktop_installer/test/network/hidden_wifi_model_test.mocks.dart
@@ -1,5 +1,5 @@
 // Mocks generated by Mockito 5.3.2 from annotations
-// in ubuntu_desktop_installer/test/connect_to_internet/ethernet_model_test.dart.
+// in ubuntu_desktop_installer/test/network/hidden_wifi_model_test.dart.
 // Do not manually edit this file.
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
@@ -63,6 +63,74 @@ class _FakeNetworkManagerActiveConnection_3 extends _i1.SmartFake
           parent,
           parentInvocation,
         );
+}
+
+/// A class which mocks [NetworkManagerAccessPoint].
+///
+/// See the documentation for Mockito's code generation for more information.
+class MockNetworkManagerAccessPoint extends _i1.Mock
+    implements _i2.NetworkManagerAccessPoint {
+  MockNetworkManagerAccessPoint() {
+    _i1.throwOnMissingStub(this);
+  }
+
+  @override
+  _i3.Stream<List<String>> get propertiesChanged => (super.noSuchMethod(
+        Invocation.getter(#propertiesChanged),
+        returnValue: _i3.Stream<List<String>>.empty(),
+      ) as _i3.Stream<List<String>>);
+  @override
+  List<_i2.NetworkManagerWifiAccessPointFlag> get flags => (super.noSuchMethod(
+        Invocation.getter(#flags),
+        returnValue: <_i2.NetworkManagerWifiAccessPointFlag>[],
+      ) as List<_i2.NetworkManagerWifiAccessPointFlag>);
+  @override
+  List<_i2.NetworkManagerWifiAccessPointSecurityFlag> get wpaFlags =>
+      (super.noSuchMethod(
+        Invocation.getter(#wpaFlags),
+        returnValue: <_i2.NetworkManagerWifiAccessPointSecurityFlag>[],
+      ) as List<_i2.NetworkManagerWifiAccessPointSecurityFlag>);
+  @override
+  List<_i2.NetworkManagerWifiAccessPointSecurityFlag> get rsnFlags =>
+      (super.noSuchMethod(
+        Invocation.getter(#rsnFlags),
+        returnValue: <_i2.NetworkManagerWifiAccessPointSecurityFlag>[],
+      ) as List<_i2.NetworkManagerWifiAccessPointSecurityFlag>);
+  @override
+  List<int> get ssid => (super.noSuchMethod(
+        Invocation.getter(#ssid),
+        returnValue: <int>[],
+      ) as List<int>);
+  @override
+  int get frequency => (super.noSuchMethod(
+        Invocation.getter(#frequency),
+        returnValue: 0,
+      ) as int);
+  @override
+  String get hwAddress => (super.noSuchMethod(
+        Invocation.getter(#hwAddress),
+        returnValue: '',
+      ) as String);
+  @override
+  _i2.NetworkManagerWifiMode get mode => (super.noSuchMethod(
+        Invocation.getter(#mode),
+        returnValue: _i2.NetworkManagerWifiMode.unknown,
+      ) as _i2.NetworkManagerWifiMode);
+  @override
+  int get maxBitrate => (super.noSuchMethod(
+        Invocation.getter(#maxBitrate),
+        returnValue: 0,
+      ) as int);
+  @override
+  int get strength => (super.noSuchMethod(
+        Invocation.getter(#strength),
+        returnValue: 0,
+      ) as int);
+  @override
+  int get lastSeen => (super.noSuchMethod(
+        Invocation.getter(#lastSeen),
+        returnValue: 0,
+      ) as int);
 }
 
 /// A class which mocks [NetworkManagerActiveConnection].
@@ -305,6 +373,63 @@ class MockNetworkManagerDevice extends _i1.Mock
         returnValue: _i3.Future<void>.value(),
         returnValueForMissingStub: _i3.Future<void>.value(),
       ) as _i3.Future<void>);
+}
+
+/// A class which mocks [NetworkManagerDeviceWireless].
+///
+/// See the documentation for Mockito's code generation for more information.
+class MockNetworkManagerDeviceWireless extends _i1.Mock
+    implements _i2.NetworkManagerDeviceWireless {
+  MockNetworkManagerDeviceWireless() {
+    _i1.throwOnMissingStub(this);
+  }
+
+  @override
+  _i3.Stream<List<String>> get propertiesChanged => (super.noSuchMethod(
+        Invocation.getter(#propertiesChanged),
+        returnValue: _i3.Stream<List<String>>.empty(),
+      ) as _i3.Stream<List<String>>);
+  @override
+  String get permHwAddress => (super.noSuchMethod(
+        Invocation.getter(#permHwAddress),
+        returnValue: '',
+      ) as String);
+  @override
+  _i2.NetworkManagerWifiMode get mode => (super.noSuchMethod(
+        Invocation.getter(#mode),
+        returnValue: _i2.NetworkManagerWifiMode.unknown,
+      ) as _i2.NetworkManagerWifiMode);
+  @override
+  int get bitrate => (super.noSuchMethod(
+        Invocation.getter(#bitrate),
+        returnValue: 0,
+      ) as int);
+  @override
+  List<_i2.NetworkManagerAccessPoint> get accessPoints => (super.noSuchMethod(
+        Invocation.getter(#accessPoints),
+        returnValue: <_i2.NetworkManagerAccessPoint>[],
+      ) as List<_i2.NetworkManagerAccessPoint>);
+  @override
+  List<_i2.NetworkManagerDeviceWifiCapability> get wirelessCapabilities =>
+      (super.noSuchMethod(
+        Invocation.getter(#wirelessCapabilities),
+        returnValue: <_i2.NetworkManagerDeviceWifiCapability>[],
+      ) as List<_i2.NetworkManagerDeviceWifiCapability>);
+  @override
+  int get lastScan => (super.noSuchMethod(
+        Invocation.getter(#lastScan),
+        returnValue: 0,
+      ) as int);
+  @override
+  _i3.Future<dynamic> requestScan({List<List<int>>? ssids}) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #requestScan,
+          [],
+          {#ssids: ssids},
+        ),
+        returnValue: _i3.Future<dynamic>.value(),
+      ) as _i3.Future<dynamic>);
 }
 
 /// A class which mocks [NetworkManagerSettingsConnection].
