+++ packages/ubuntu_desktop_installer/test/network/hidden_wifi_view_test.mocks.dart
@@ -1,16 +1,17 @@
 // Mocks generated by Mockito 5.3.2 from annotations
-// in ubuntu_desktop_installer/test/connect_to_internet/ethernet_view_test.dart.
+// in ubuntu_desktop_installer/test/network/hidden_wifi_view_test.dart.
 // Do not manually edit this file.
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
-import 'dart:async' as _i5;
-import 'dart:ui' as _i6;
+import 'dart:async' as _i6;
+import 'dart:ui' as _i7;
 
 import 'package:mockito/mockito.dart' as _i1;
-import 'package:ubuntu_desktop_installer/pages/connect_to_internet/connect_model.dart'
+import 'package:ubuntu_desktop_installer/pages/network/connect_model.dart'
+    as _i5;
+import 'package:ubuntu_desktop_installer/pages/network/hidden_wifi_model.dart'
     as _i4;
-import 'package:ubuntu_desktop_installer/pages/connect_to_internet/ethernet_model.dart'
-    as _i3;
+import 'package:ubuntu_desktop_installer/pages/network/wifi_model.dart' as _i3;
 import 'package:ubuntu_desktop_installer/services.dart' as _i2;
 
 // ignore_for_file: type=lint
@@ -35,9 +36,8 @@ class _FakeNetworkService_0 extends _i1.SmartFake
         );
 }
 
-class _FakeEthernetDevice_1 extends _i1.SmartFake
-    implements _i3.EthernetDevice {
-  _FakeEthernetDevice_1(
+class _FakeWifiDevice_1 extends _i1.SmartFake implements _i3.WifiDevice {
+  _FakeWifiDevice_1(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -57,11 +57,11 @@ class _FakeNetworkManagerDevice_2 extends _i1.SmartFake
         );
 }
 
-/// A class which mocks [EthernetModel].
+/// A class which mocks [HiddenWifiModel].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockEthernetModel extends _i1.Mock implements _i3.EthernetModel {
-  MockEthernetModel() {
+class MockHiddenWifiModel extends _i1.Mock implements _i4.HiddenWifiModel {
+  MockHiddenWifiModel() {
     _i1.throwOnMissingStub(this);
   }
 
@@ -76,13 +76,13 @@ class MockEthernetModel extends _i1.Mock implements _i3.EthernetModel {
         returnValue: false,
       ) as bool);
   @override
-  bool get isConnecting => (super.noSuchMethod(
-        Invocation.getter(#isConnecting),
+  bool get hasActiveConnection => (super.noSuchMethod(
+        Invocation.getter(#hasActiveConnection),
         returnValue: false,
       ) as bool);
   @override
-  bool get hasActiveConnection => (super.noSuchMethod(
-        Invocation.getter(#hasActiveConnection),
+  bool get isConnecting => (super.noSuchMethod(
+        Invocation.getter(#isConnecting),
         returnValue: false,
       ) as bool);
   @override
@@ -91,10 +91,15 @@ class MockEthernetModel extends _i1.Mock implements _i3.EthernetModel {
         returnValue: false,
       ) as bool);
   @override
-  _i4.ConnectMode get connectMode => (super.noSuchMethod(
+  _i5.ConnectMode get connectMode => (super.noSuchMethod(
         Invocation.getter(#connectMode),
-        returnValue: _i4.ConnectMode.none,
-      ) as _i4.ConnectMode);
+        returnValue: _i5.ConnectMode.none,
+      ) as _i5.ConnectMode);
+  @override
+  String get ssid => (super.noSuchMethod(
+        Invocation.getter(#ssid),
+        returnValue: '',
+      ) as String);
   @override
   _i2.NetworkService get service => (super.noSuchMethod(
         Invocation.getter(#service),
@@ -104,15 +109,15 @@ class MockEthernetModel extends _i1.Mock implements _i3.EthernetModel {
         ),
       ) as _i2.NetworkService);
   @override
-  List<_i3.EthernetDevice> get devices => (super.noSuchMethod(
+  List<_i3.WifiDevice> get devices => (super.noSuchMethod(
         Invocation.getter(#devices),
-        returnValue: <_i3.EthernetDevice>[],
-      ) as List<_i3.EthernetDevice>);
+        returnValue: <_i3.WifiDevice>[],
+      ) as List<_i3.WifiDevice>);
   @override
-  _i5.Stream<dynamic> get onAvailabilityChanged => (super.noSuchMethod(
+  _i6.Stream<dynamic> get onAvailabilityChanged => (super.noSuchMethod(
         Invocation.getter(#onAvailabilityChanged),
-        returnValue: _i5.Stream<dynamic>.empty(),
-      ) as _i5.Stream<dynamic>);
+        returnValue: _i6.Stream<dynamic>.empty(),
+      ) as _i6.Stream<dynamic>);
   @override
   bool get hasListeners => (super.noSuchMethod(
         Invocation.getter(#hasListeners),
@@ -124,13 +129,30 @@ class MockEthernetModel extends _i1.Mock implements _i3.EthernetModel {
         returnValue: false,
       ) as bool);
   @override
-  void onSelected() => super.noSuchMethod(
+  _i6.Future<dynamic> connect() => (super.noSuchMethod(
         Invocation.method(
-          #onSelected,
+          #connect,
           [],
+        ),
+        returnValue: _i6.Future<dynamic>.value(),
+      ) as _i6.Future<dynamic>);
+  @override
+  void setSsid(String? ssid) => super.noSuchMethod(
+        Invocation.method(
+          #setSsid,
+          [ssid],
         ),
         returnValueForMissingStub: null,
       );
+  @override
+  _i6.Future<void> enable() => (super.noSuchMethod(
+        Invocation.method(
+          #enable,
+          [],
+        ),
+        returnValue: _i6.Future<void>.value(),
+        returnValueForMissingStub: _i6.Future<void>.value(),
+      ) as _i6.Future<void>);
   @override
   void onDeselected() => super.noSuchMethod(
         Invocation.method(
@@ -140,23 +162,13 @@ class MockEthernetModel extends _i1.Mock implements _i3.EthernetModel {
         returnValueForMissingStub: null,
       );
   @override
-  _i5.Future<void> enable() => (super.noSuchMethod(
+  void onSelected() => super.noSuchMethod(
         Invocation.method(
-          #enable,
+          #onSelected,
           [],
         ),
-        returnValue: _i5.Future<void>.value(),
-        returnValueForMissingStub: _i5.Future<void>.value(),
-      ) as _i5.Future<void>);
-  @override
-  _i5.Future<void> connect() => (super.noSuchMethod(
-        Invocation.method(
-          #connect,
-          [],
-        ),
-        returnValue: _i5.Future<void>.value(),
-        returnValueForMissingStub: _i5.Future<void>.value(),
-      ) as _i5.Future<void>);
+        returnValueForMissingStub: null,
+      );
   @override
   List<_i2.NetworkManagerDevice> getDevices() => (super.noSuchMethod(
         Invocation.method(
@@ -166,38 +178,38 @@ class MockEthernetModel extends _i1.Mock implements _i3.EthernetModel {
         returnValue: <_i2.NetworkManagerDevice>[],
       ) as List<_i2.NetworkManagerDevice>);
   @override
-  _i3.EthernetDevice createDevice(_i2.NetworkManagerDevice? device) =>
+  _i3.WifiDevice createDevice(_i2.NetworkManagerDevice? device) =>
       (super.noSuchMethod(
         Invocation.method(
           #createDevice,
           [device],
         ),
-        returnValue: _FakeEthernetDevice_1(
+        returnValue: _FakeWifiDevice_1(
           this,
           Invocation.method(
             #createDevice,
             [device],
           ),
         ),
-      ) as _i3.EthernetDevice);
+      ) as _i3.WifiDevice);
   @override
-  _i5.Future<void> init() => (super.noSuchMethod(
+  _i6.Future<void> init() => (super.noSuchMethod(
         Invocation.method(
           #init,
           [],
         ),
-        returnValue: _i5.Future<void>.value(),
-        returnValueForMissingStub: _i5.Future<void>.value(),
-      ) as _i5.Future<void>);
+        returnValue: _i6.Future<void>.value(),
+        returnValueForMissingStub: _i6.Future<void>.value(),
+      ) as _i6.Future<void>);
   @override
-  _i5.Future<void> cleanup() => (super.noSuchMethod(
+  _i6.Future<void> cleanup() => (super.noSuchMethod(
         Invocation.method(
           #cleanup,
           [],
         ),
-        returnValue: _i5.Future<void>.value(),
-        returnValueForMissingStub: _i5.Future<void>.value(),
-      ) as _i5.Future<void>);
+        returnValue: _i6.Future<void>.value(),
+        returnValueForMissingStub: _i6.Future<void>.value(),
+      ) as _i6.Future<void>);
   @override
   void dispose() => super.noSuchMethod(
         Invocation.method(
@@ -215,7 +227,7 @@ class MockEthernetModel extends _i1.Mock implements _i3.EthernetModel {
         returnValueForMissingStub: null,
       );
   @override
-  bool isSelectedDevice(_i3.EthernetDevice? device) => (super.noSuchMethod(
+  bool isSelectedDevice(_i3.WifiDevice? device) => (super.noSuchMethod(
         Invocation.method(
           #isSelectedDevice,
           [device],
@@ -223,7 +235,7 @@ class MockEthernetModel extends _i1.Mock implements _i3.EthernetModel {
         returnValue: false,
       ) as bool);
   @override
-  void selectDevice(_i3.EthernetDevice? device) => super.noSuchMethod(
+  void selectDevice(_i3.WifiDevice? device) => super.noSuchMethod(
         Invocation.method(
           #selectDevice,
           [device],
@@ -239,7 +251,7 @@ class MockEthernetModel extends _i1.Mock implements _i3.EthernetModel {
         returnValueForMissingStub: null,
       );
   @override
-  void setProperties(_i5.Stream<List<String>>? properties) =>
+  void setProperties(_i6.Stream<List<String>>? properties) =>
       super.noSuchMethod(
         Invocation.method(
           #setProperties,
@@ -250,7 +262,7 @@ class MockEthernetModel extends _i1.Mock implements _i3.EthernetModel {
   @override
   void addPropertyListener(
     String? property,
-    _i6.VoidCallback? onChanged,
+    _i7.VoidCallback? onChanged,
   ) =>
       super.noSuchMethod(
         Invocation.method(
@@ -279,7 +291,7 @@ class MockEthernetModel extends _i1.Mock implements _i3.EthernetModel {
         returnValueForMissingStub: null,
       );
   @override
-  void addListener(_i6.VoidCallback? listener) => super.noSuchMethod(
+  void addListener(_i7.VoidCallback? listener) => super.noSuchMethod(
         Invocation.method(
           #addListener,
           [listener],
@@ -287,7 +299,7 @@ class MockEthernetModel extends _i1.Mock implements _i3.EthernetModel {
         returnValueForMissingStub: null,
       );
   @override
-  void removeListener(_i6.VoidCallback? listener) => super.noSuchMethod(
+  void removeListener(_i7.VoidCallback? listener) => super.noSuchMethod(
         Invocation.method(
           #removeListener,
           [listener],
@@ -296,14 +308,34 @@ class MockEthernetModel extends _i1.Mock implements _i3.EthernetModel {
       );
 }
 
-/// A class which mocks [EthernetDevice].
+/// A class which mocks [WifiDevice].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockEthernetDevice extends _i1.Mock implements _i3.EthernetDevice {
-  MockEthernetDevice() {
+class MockWifiDevice extends _i1.Mock implements _i3.WifiDevice {
+  MockWifiDevice() {
     _i1.throwOnMissingStub(this);
   }
 
+  @override
+  bool get isActive => (super.noSuchMethod(
+        Invocation.getter(#isActive),
+        returnValue: false,
+      ) as bool);
+  @override
+  List<_i3.AccessPoint> get accessPoints => (super.noSuchMethod(
+        Invocation.getter(#accessPoints),
+        returnValue: <_i3.AccessPoint>[],
+      ) as List<_i3.AccessPoint>);
+  @override
+  bool get scanning => (super.noSuchMethod(
+        Invocation.getter(#scanning),
+        returnValue: false,
+      ) as bool);
+  @override
+  int get lastScan => (super.noSuchMethod(
+        Invocation.getter(#lastScan),
+        returnValue: 0,
+      ) as int);
   @override
   _i2.NetworkManagerDevice get device => (super.noSuchMethod(
         Invocation.getter(#device),
@@ -322,11 +354,6 @@ class MockEthernetDevice extends _i1.Mock implements _i3.EthernetDevice {
         Invocation.getter(#state),
         returnValue: _i2.NetworkManagerDeviceState.unknown,
       ) as _i2.NetworkManagerDeviceState);
-  @override
-  bool get isActive => (super.noSuchMethod(
-        Invocation.getter(#isActive),
-        returnValue: false,
-      ) as bool);
   @override
   bool get isConnecting => (super.noSuchMethod(
         Invocation.getter(#isConnecting),
@@ -385,6 +412,14 @@ class MockEthernetDevice extends _i1.Mock implements _i3.EthernetDevice {
         returnValueForMissingStub: null,
       );
   @override
+  void dispose() => super.noSuchMethod(
+        Invocation.method(
+          #dispose,
+          [],
+        ),
+        returnValueForMissingStub: null,
+      );
+  @override
   void updateDevice(_i2.NetworkManagerDevice? device) => super.noSuchMethod(
         Invocation.method(
           #updateDevice,
@@ -393,16 +428,57 @@ class MockEthernetDevice extends _i1.Mock implements _i3.EthernetDevice {
         returnValueForMissingStub: null,
       );
   @override
-  _i5.Future<void> disconnect() => (super.noSuchMethod(
+  bool isActiveAccessPoint(_i3.AccessPoint? accessPoint) => (super.noSuchMethod(
+        Invocation.method(
+          #isActiveAccessPoint,
+          [accessPoint],
+        ),
+        returnValue: false,
+      ) as bool);
+  @override
+  bool isSelectedAccessPoint(_i3.AccessPoint? accessPoint) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #isSelectedAccessPoint,
+          [accessPoint],
+        ),
+        returnValue: false,
+      ) as bool);
+  @override
+  void selectAccessPoint(_i3.AccessPoint? accessPoint) => super.noSuchMethod(
+        Invocation.method(
+          #selectAccessPoint,
+          [accessPoint],
+        ),
+        returnValueForMissingStub: null,
+      );
+  @override
+  _i6.Future<void> requestScan({String? ssid}) => (super.noSuchMethod(
+        Invocation.method(
+          #requestScan,
+          [],
+          {#ssid: ssid},
+        ),
+        returnValue: _i6.Future<void>.value(),
+        returnValueForMissingStub: _i6.Future<void>.value(),
+      ) as _i6.Future<void>);
+  @override
+  _i3.AccessPoint? findAccessPoint(String? ssid) =>
+      (super.noSuchMethod(Invocation.method(
+        #findAccessPoint,
+        [ssid],
+      )) as _i3.AccessPoint?);
+  @override
+  _i6.Future<void> disconnect() => (super.noSuchMethod(
         Invocation.method(
           #disconnect,
           [],
         ),
-        returnValue: _i5.Future<void>.value(),
-        returnValueForMissingStub: _i5.Future<void>.value(),
-      ) as _i5.Future<void>);
+        returnValue: _i6.Future<void>.value(),
+        returnValueForMissingStub: _i6.Future<void>.value(),
+      ) as _i6.Future<void>);
   @override
-  void setProperties(_i5.Stream<List<String>>? properties) =>
+  void setProperties(_i6.Stream<List<String>>? properties) =>
       super.noSuchMethod(
         Invocation.method(
           #setProperties,
@@ -413,7 +489,7 @@ class MockEthernetDevice extends _i1.Mock implements _i3.EthernetDevice {
   @override
   void addPropertyListener(
     String? property,
-    _i6.VoidCallback? onChanged,
+    _i7.VoidCallback? onChanged,
   ) =>
       super.noSuchMethod(
         Invocation.method(
@@ -442,15 +518,7 @@ class MockEthernetDevice extends _i1.Mock implements _i3.EthernetDevice {
         returnValueForMissingStub: null,
       );
   @override
-  void dispose() => super.noSuchMethod(
-        Invocation.method(
-          #dispose,
-          [],
-        ),
-        returnValueForMissingStub: null,
-      );
-  @override
-  void addListener(_i6.VoidCallback? listener) => super.noSuchMethod(
+  void addListener(_i7.VoidCallback? listener) => super.noSuchMethod(
         Invocation.method(
           #addListener,
           [listener],
@@ -458,7 +526,7 @@ class MockEthernetDevice extends _i1.Mock implements _i3.EthernetDevice {
         returnValueForMissingStub: null,
       );
   @override
-  void removeListener(_i6.VoidCallback? listener) => super.noSuchMethod(
+  void removeListener(_i7.VoidCallback? listener) => super.noSuchMethod(
         Invocation.method(
           #removeListener,
           [listener],
