+++ packages/ubuntu_desktop_installer/test/network/network_device_test.dart
@@ -3,10 +3,10 @@ import 'dart:async';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/annotations.dart';
 import 'package:mockito/mockito.dart';
-import 'package:ubuntu_desktop_installer/pages/connect_to_internet/network_model.dart';
+import 'package:ubuntu_desktop_installer/pages/network/network_device.dart';
 import 'package:ubuntu_desktop_installer/services.dart';
 
-import 'network_model_test.mocks.dart';
+import 'network_device_test.mocks.dart';
 
 @GenerateMocks([
   NetworkManagerActiveConnection,
