+++ packages/ubuntu_desktop_installer/test/network/network_model_test.mocks.dart
@@ -1,5 +1,5 @@
 // Mocks generated by Mockito 5.3.2 from annotations
-// in ubuntu_desktop_installer/test/connect_to_internet/connect_to_internet_model_test.dart.
+// in ubuntu_desktop_installer/test/network/network_model_test.dart.
 // Do not manually edit this file.
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
@@ -7,7 +7,7 @@ import 'dart:async' as _i3;
 import 'dart:ui' as _i4;
 
 import 'package:mockito/mockito.dart' as _i1;
-import 'package:ubuntu_desktop_installer/pages/connect_to_internet/connect_model.dart'
+import 'package:ubuntu_desktop_installer/pages/network/connect_model.dart'
     as _i2;
 
 // ignore_for_file: type=lint
