+++ packages/ubuntu_desktop_installer/test/network/network_page_test.mocks.dart
@@ -1,5 +1,5 @@
 // Mocks generated by Mockito 5.3.2 from annotations
-// in ubuntu_desktop_installer/test/connect_to_internet/connect_to_internet_page_test.dart.
+// in ubuntu_desktop_installer/test/network/network_page_test.dart.
 // Do not manually edit this file.
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
@@ -8,16 +8,15 @@ import 'dart:ui' as _i6;
 
 import 'package:dbus/dbus.dart' as _i10;
 import 'package:mockito/mockito.dart' as _i1;
-import 'package:ubuntu_desktop_installer/pages/connect_to_internet/connect_model.dart'
-    as _i8;
-import 'package:ubuntu_desktop_installer/pages/connect_to_internet/connect_to_internet_model.dart'
+import 'package:ubuntu_desktop_installer/pages/network/connect_model.dart'
     as _i7;
-import 'package:ubuntu_desktop_installer/pages/connect_to_internet/ethernet_model.dart'
+import 'package:ubuntu_desktop_installer/pages/network/ethernet_model.dart'
     as _i3;
-import 'package:ubuntu_desktop_installer/pages/connect_to_internet/hidden_wifi_model.dart'
+import 'package:ubuntu_desktop_installer/pages/network/hidden_wifi_model.dart'
+    as _i8;
+import 'package:ubuntu_desktop_installer/pages/network/network_model.dart'
     as _i9;
-import 'package:ubuntu_desktop_installer/pages/connect_to_internet/wifi_model.dart'
-    as _i4;
+import 'package:ubuntu_desktop_installer/pages/network/wifi_model.dart' as _i4;
 import 'package:ubuntu_desktop_installer/services.dart' as _i2;
 
 // ignore_for_file: type=lint
@@ -270,171 +269,6 @@ class MockAccessPoint extends _i1.Mock implements _i4.AccessPoint {
       );
 }
 
-/// A class which mocks [ConnectToInternetModel].
-///
-/// See the documentation for Mockito's code generation for more information.
-class MockConnectToInternetModel extends _i1.Mock
-    implements _i7.ConnectToInternetModel {
-  MockConnectToInternetModel() {
-    _i1.throwOnMissingStub(this);
-  }
-
-  @override
-  bool get canConnect => (super.noSuchMethod(
-        Invocation.getter(#canConnect),
-        returnValue: false,
-      ) as bool);
-  @override
-  bool get isConnected => (super.noSuchMethod(
-        Invocation.getter(#isConnected),
-        returnValue: false,
-      ) as bool);
-  @override
-  bool get hasActiveConnection => (super.noSuchMethod(
-        Invocation.getter(#hasActiveConnection),
-        returnValue: false,
-      ) as bool);
-  @override
-  bool get isConnecting => (super.noSuchMethod(
-        Invocation.getter(#isConnecting),
-        returnValue: false,
-      ) as bool);
-  @override
-  bool get isEnabled => (super.noSuchMethod(
-        Invocation.getter(#isEnabled),
-        returnValue: false,
-      ) as bool);
-  @override
-  _i5.Stream<dynamic> get onAvailabilityChanged => (super.noSuchMethod(
-        Invocation.getter(#onAvailabilityChanged),
-        returnValue: _i5.Stream<dynamic>.empty(),
-      ) as _i5.Stream<dynamic>);
-  @override
-  _i8.ConnectMode get connectMode => (super.noSuchMethod(
-        Invocation.getter(#connectMode),
-        returnValue: _i8.ConnectMode.none,
-      ) as _i8.ConnectMode);
-  @override
-  bool get hasListeners => (super.noSuchMethod(
-        Invocation.getter(#hasListeners),
-        returnValue: false,
-      ) as bool);
-  @override
-  bool get isDisposed => (super.noSuchMethod(
-        Invocation.getter(#isDisposed),
-        returnValue: false,
-      ) as bool);
-  @override
-  void addConnectMode(_i8.ConnectModel? model) => super.noSuchMethod(
-        Invocation.method(
-          #addConnectMode,
-          [model],
-        ),
-        returnValueForMissingStub: null,
-      );
-  @override
-  void selectConnectMode([_i8.ConnectMode? mode]) => super.noSuchMethod(
-        Invocation.method(
-          #selectConnectMode,
-          [mode],
-        ),
-        returnValueForMissingStub: null,
-      );
-  @override
-  void onSelected() => super.noSuchMethod(
-        Invocation.method(
-          #onSelected,
-          [],
-        ),
-        returnValueForMissingStub: null,
-      );
-  @override
-  void onDeselected() => super.noSuchMethod(
-        Invocation.method(
-          #onDeselected,
-          [],
-        ),
-        returnValueForMissingStub: null,
-      );
-  @override
-  _i5.Future<void> init() => (super.noSuchMethod(
-        Invocation.method(
-          #init,
-          [],
-        ),
-        returnValue: _i5.Future<void>.value(),
-        returnValueForMissingStub: _i5.Future<void>.value(),
-      ) as _i5.Future<void>);
-  @override
-  _i5.Future<void> markConfigured() => (super.noSuchMethod(
-        Invocation.method(
-          #markConfigured,
-          [],
-        ),
-        returnValue: _i5.Future<void>.value(),
-        returnValueForMissingStub: _i5.Future<void>.value(),
-      ) as _i5.Future<void>);
-  @override
-  _i5.Future<void> cleanup() => (super.noSuchMethod(
-        Invocation.method(
-          #cleanup,
-          [],
-        ),
-        returnValue: _i5.Future<void>.value(),
-        returnValueForMissingStub: _i5.Future<void>.value(),
-      ) as _i5.Future<void>);
-  @override
-  _i5.Future<void> enable() => (super.noSuchMethod(
-        Invocation.method(
-          #enable,
-          [],
-        ),
-        returnValue: _i5.Future<void>.value(),
-        returnValueForMissingStub: _i5.Future<void>.value(),
-      ) as _i5.Future<void>);
-  @override
-  _i5.Future<void> connect() => (super.noSuchMethod(
-        Invocation.method(
-          #connect,
-          [],
-        ),
-        returnValue: _i5.Future<void>.value(),
-        returnValueForMissingStub: _i5.Future<void>.value(),
-      ) as _i5.Future<void>);
-  @override
-  void addListener(_i6.VoidCallback? listener) => super.noSuchMethod(
-        Invocation.method(
-          #addListener,
-          [listener],
-        ),
-        returnValueForMissingStub: null,
-      );
-  @override
-  void removeListener(_i6.VoidCallback? listener) => super.noSuchMethod(
-        Invocation.method(
-          #removeListener,
-          [listener],
-        ),
-        returnValueForMissingStub: null,
-      );
-  @override
-  void dispose() => super.noSuchMethod(
-        Invocation.method(
-          #dispose,
-          [],
-        ),
-        returnValueForMissingStub: null,
-      );
-  @override
-  void notifyListeners() => super.noSuchMethod(
-        Invocation.method(
-          #notifyListeners,
-          [],
-        ),
-        returnValueForMissingStub: null,
-      );
-}
-
 /// A class which mocks [EthernetModel].
 ///
 /// See the documentation for Mockito's code generation for more information.
@@ -469,10 +303,10 @@ class MockEthernetModel extends _i1.Mock implements _i3.EthernetModel {
         returnValue: false,
       ) as bool);
   @override
-  _i8.ConnectMode get connectMode => (super.noSuchMethod(
+  _i7.ConnectMode get connectMode => (super.noSuchMethod(
         Invocation.getter(#connectMode),
-        returnValue: _i8.ConnectMode.none,
-      ) as _i8.ConnectMode);
+        returnValue: _i7.ConnectMode.none,
+      ) as _i7.ConnectMode);
   @override
   _i2.NetworkService get service => (super.noSuchMethod(
         Invocation.getter(#service),
@@ -856,7 +690,7 @@ class MockEthernetDevice extends _i1.Mock implements _i3.EthernetDevice {
 /// A class which mocks [HiddenWifiModel].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockHiddenWifiModel extends _i1.Mock implements _i9.HiddenWifiModel {
+class MockHiddenWifiModel extends _i1.Mock implements _i8.HiddenWifiModel {
   MockHiddenWifiModel() {
     _i1.throwOnMissingStub(this);
   }
@@ -887,10 +721,10 @@ class MockHiddenWifiModel extends _i1.Mock implements _i9.HiddenWifiModel {
         returnValue: false,
       ) as bool);
   @override
-  _i8.ConnectMode get connectMode => (super.noSuchMethod(
+  _i7.ConnectMode get connectMode => (super.noSuchMethod(
         Invocation.getter(#connectMode),
-        returnValue: _i8.ConnectMode.none,
-      ) as _i8.ConnectMode);
+        returnValue: _i7.ConnectMode.none,
+      ) as _i7.ConnectMode);
   @override
   String get ssid => (super.noSuchMethod(
         Invocation.getter(#ssid),
@@ -1099,6 +933,170 @@ class MockHiddenWifiModel extends _i1.Mock implements _i9.HiddenWifiModel {
         Invocation.method(
           #removeListener,
           [listener],
+        ),
+        returnValueForMissingStub: null,
+      );
+}
+
+/// A class which mocks [NetworkModel].
+///
+/// See the documentation for Mockito's code generation for more information.
+class MockNetworkModel extends _i1.Mock implements _i9.NetworkModel {
+  MockNetworkModel() {
+    _i1.throwOnMissingStub(this);
+  }
+
+  @override
+  bool get canConnect => (super.noSuchMethod(
+        Invocation.getter(#canConnect),
+        returnValue: false,
+      ) as bool);
+  @override
+  bool get isConnected => (super.noSuchMethod(
+        Invocation.getter(#isConnected),
+        returnValue: false,
+      ) as bool);
+  @override
+  bool get hasActiveConnection => (super.noSuchMethod(
+        Invocation.getter(#hasActiveConnection),
+        returnValue: false,
+      ) as bool);
+  @override
+  bool get isConnecting => (super.noSuchMethod(
+        Invocation.getter(#isConnecting),
+        returnValue: false,
+      ) as bool);
+  @override
+  bool get isEnabled => (super.noSuchMethod(
+        Invocation.getter(#isEnabled),
+        returnValue: false,
+      ) as bool);
+  @override
+  _i5.Stream<dynamic> get onAvailabilityChanged => (super.noSuchMethod(
+        Invocation.getter(#onAvailabilityChanged),
+        returnValue: _i5.Stream<dynamic>.empty(),
+      ) as _i5.Stream<dynamic>);
+  @override
+  _i7.ConnectMode get connectMode => (super.noSuchMethod(
+        Invocation.getter(#connectMode),
+        returnValue: _i7.ConnectMode.none,
+      ) as _i7.ConnectMode);
+  @override
+  bool get hasListeners => (super.noSuchMethod(
+        Invocation.getter(#hasListeners),
+        returnValue: false,
+      ) as bool);
+  @override
+  bool get isDisposed => (super.noSuchMethod(
+        Invocation.getter(#isDisposed),
+        returnValue: false,
+      ) as bool);
+  @override
+  void addConnectMode(_i7.ConnectModel? model) => super.noSuchMethod(
+        Invocation.method(
+          #addConnectMode,
+          [model],
+        ),
+        returnValueForMissingStub: null,
+      );
+  @override
+  void selectConnectMode([_i7.ConnectMode? mode]) => super.noSuchMethod(
+        Invocation.method(
+          #selectConnectMode,
+          [mode],
+        ),
+        returnValueForMissingStub: null,
+      );
+  @override
+  void onSelected() => super.noSuchMethod(
+        Invocation.method(
+          #onSelected,
+          [],
+        ),
+        returnValueForMissingStub: null,
+      );
+  @override
+  void onDeselected() => super.noSuchMethod(
+        Invocation.method(
+          #onDeselected,
+          [],
+        ),
+        returnValueForMissingStub: null,
+      );
+  @override
+  _i5.Future<void> init() => (super.noSuchMethod(
+        Invocation.method(
+          #init,
+          [],
+        ),
+        returnValue: _i5.Future<void>.value(),
+        returnValueForMissingStub: _i5.Future<void>.value(),
+      ) as _i5.Future<void>);
+  @override
+  _i5.Future<void> markConfigured() => (super.noSuchMethod(
+        Invocation.method(
+          #markConfigured,
+          [],
+        ),
+        returnValue: _i5.Future<void>.value(),
+        returnValueForMissingStub: _i5.Future<void>.value(),
+      ) as _i5.Future<void>);
+  @override
+  _i5.Future<void> cleanup() => (super.noSuchMethod(
+        Invocation.method(
+          #cleanup,
+          [],
+        ),
+        returnValue: _i5.Future<void>.value(),
+        returnValueForMissingStub: _i5.Future<void>.value(),
+      ) as _i5.Future<void>);
+  @override
+  _i5.Future<void> enable() => (super.noSuchMethod(
+        Invocation.method(
+          #enable,
+          [],
+        ),
+        returnValue: _i5.Future<void>.value(),
+        returnValueForMissingStub: _i5.Future<void>.value(),
+      ) as _i5.Future<void>);
+  @override
+  _i5.Future<void> connect() => (super.noSuchMethod(
+        Invocation.method(
+          #connect,
+          [],
+        ),
+        returnValue: _i5.Future<void>.value(),
+        returnValueForMissingStub: _i5.Future<void>.value(),
+      ) as _i5.Future<void>);
+  @override
+  void addListener(_i6.VoidCallback? listener) => super.noSuchMethod(
+        Invocation.method(
+          #addListener,
+          [listener],
+        ),
+        returnValueForMissingStub: null,
+      );
+  @override
+  void removeListener(_i6.VoidCallback? listener) => super.noSuchMethod(
+        Invocation.method(
+          #removeListener,
+          [listener],
+        ),
+        returnValueForMissingStub: null,
+      );
+  @override
+  void dispose() => super.noSuchMethod(
+        Invocation.method(
+          #dispose,
+          [],
+        ),
+        returnValueForMissingStub: null,
+      );
+  @override
+  void notifyListeners() => super.noSuchMethod(
+        Invocation.method(
+          #notifyListeners,
+          [],
         ),
         returnValueForMissingStub: null,
       );
@@ -1487,10 +1485,10 @@ class MockWifiModel extends _i1.Mock implements _i4.WifiModel {
         returnValue: false,
       ) as bool);
   @override
-  _i8.ConnectMode get connectMode => (super.noSuchMethod(
+  _i7.ConnectMode get connectMode => (super.noSuchMethod(
         Invocation.getter(#connectMode),
-        returnValue: _i8.ConnectMode.none,
-      ) as _i8.ConnectMode);
+        returnValue: _i7.ConnectMode.none,
+      ) as _i7.ConnectMode);
   @override
   _i2.NetworkService get service => (super.noSuchMethod(
         Invocation.getter(#service),
