+++ packages/ubuntu_desktop_installer/test/network/wifi_model_test.mocks.dart
@@ -1,5 +1,5 @@
 // Mocks generated by Mockito 5.3.2 from annotations
-// in ubuntu_desktop_installer/test/connect_to_internet/hidden_wifi_model_test.dart.
+// in ubuntu_desktop_installer/test/network/wifi_model_test.dart.
 // Do not manually edit this file.
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
